backend/app/requirements.txt
fastapi
uvicorn[standard]
earthengine-api
numpy
Pillow
requests
python-multipart
pydantic
python-dotenv
